openeo_argoworkflows/api/pyproject.toml
[tool.poetry]
name = "openeo-argoworkflows-api"
version = "2026.3.11"
description = ""
authors = ["sean <sean.hoyal@external.eodc.eu>"]
readme = "README.md"
packages = [{include = "openeo_argoworkflows_api"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
#openeo_fastapi = "^2025.5.1"
openeo_fastapi = { git = "https://github.com/DLR-terrabyte/openeo-fastapi.git", branch = "main" }
python-multipart = "^0.0.16"
pyjwt = "2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
ipykernel = "^6.15.1"
pre-commit = "^2.20.0"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.23.0"
pytest-postgresql= ">=4.1.1"
pydoc-markdown = "4.8.2"
fakeredis = "2.23.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
